refactor: move proxyroles in separate files
